fix Day/Night mode for ceiling 22/23
Fix an issue related to not being able to change the Day/Night mode of yeelink.light.ceiling22/23.

Day mode should be 1, Night mode should be 2.

But if Night mode is on, when you change the brightness, the brightness level shown in HASS stays the same. But the light actually changes its brightness, and the brightness also changed in Xiaomi Home App. In Day mode, everything works fine, no problem.

I can't find a way to fix this.